Model moved dto and base packages from a dao package to a model package
