+++ geomagnetic-api/src/main/java/net/c7j/weather/geomagnetic/mapper/impl/ForecastDtoMapper.java
@@ -1,19 +1,14 @@
 package net.c7j.weather.geomagnetic.mapper.impl;
 
-import net.c7j.weather.geomagnetic.dao.base.ActiveType;
-import net.c7j.weather.geomagnetic.dao.dto.ForecastDto;
 import net.c7j.weather.geomagnetic.dao.entity.ForecastEntity;
 import net.c7j.weather.geomagnetic.mapper.AbstractDtoMapper;
+import net.c7j.weather.geomagnetic.model.base.ActiveType;
+import net.c7j.weather.geomagnetic.model.dto.ForecastDto;
 import net.c7j.weather.geomagnetic.service.DtoMapperChecker;
-
 import org.modelmapper.ModelMapper;
 import org.springframework.stereotype.Component;
 
-import java.time.Instant;
-import java.time.LocalDate;
-import java.time.LocalDateTime;
-import java.time.LocalTime;
-import java.time.ZoneOffset;
+import java.time.*;
 import java.util.function.BiConsumer;
 
 @Component
