+++ geomagnetic-api/src/main/java/net/c7j/weather/geomagnetic/model/dto/ForecastDto.java
@@ -1,8 +1,8 @@
-package net.c7j.weather.geomagnetic.dao.dto;
+package net.c7j.weather.geomagnetic.model.dto;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.*;
-import net.c7j.weather.geomagnetic.dao.base.IndexType;
+import net.c7j.weather.geomagnetic.model.base.IndexType;
 import net.c7j.weather.geomagnetic.util.JsonWriter;
 
 @Getter
