+++ geomagnetic-api/src/main/java/net/c7j/weather/geomagnetic/model/dto/ForecastEventWrapper.java
@@ -1,11 +1,10 @@
-package net.c7j.weather.geomagnetic.dao.dto;
-
-import java.util.Set;
-
-import javax.validation.constraints.NotEmpty;
+package net.c7j.weather.geomagnetic.model.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
+
+import javax.validation.constraints.NotEmpty;
+import java.util.Set;
 
 @Getter
 @AllArgsConstructor
