+++ geomagnetic-api/src/main/java/net/c7j/weather/geomagnetic/model/dto/ForecastResponse.java
@@ -1,4 +1,4 @@
-package net.c7j.weather.geomagnetic.dao.dto;
+package net.c7j.weather.geomagnetic.model.dto;
 
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
