+++ geomagnetic-api/src/test/java/net/c7j/weather/geomagnetic/dao/repository/ForecastRepositoryIntegrationTest.java
@@ -1,8 +1,8 @@
 package net.c7j.weather.geomagnetic.dao.repository;
 
-import net.c7j.weather.geomagnetic.dao.base.ActiveType;
-import net.c7j.weather.geomagnetic.dao.base.IndexType;
 import net.c7j.weather.geomagnetic.dao.entity.ForecastEntity;
+import net.c7j.weather.geomagnetic.model.base.ActiveType;
+import net.c7j.weather.geomagnetic.model.base.IndexType;
 import net.c7j.weather.geomagnetic.test.helper.GeneratorHelper;
 import net.c7j.weather.geomagnetic.test.tag.IntegrationTest;
 import org.junit.jupiter.api.BeforeEach;
