+++ geomagnetic-api/src/test/java/net/c7j/weather/geomagnetic/event/listener/ForecastListenerTest.java
@@ -1,7 +1,7 @@
 package net.c7j.weather.geomagnetic.event.listener;
 
 import net.c7j.weather.geomagnetic.dao.access.ForecastAccessService;
-import net.c7j.weather.geomagnetic.dao.dto.ForecastEventWrapper;
+import net.c7j.weather.geomagnetic.model.dto.ForecastEventWrapper;
 import net.c7j.weather.geomagnetic.service.ForecastUpsertService;
 import net.c7j.weather.geomagnetic.test.helper.GeneratorHelper;
 import net.c7j.weather.geomagnetic.test.tag.IntegrationTest;
