+++ geomagnetic-api/src/test/java/net/c7j/weather/geomagnetic/mapper/ForecastDtoMapperIntegrationTest.java
@@ -1,9 +1,9 @@
 package net.c7j.weather.geomagnetic.mapper;
 
-import net.c7j.weather.geomagnetic.dao.base.IndexType;
-import net.c7j.weather.geomagnetic.dao.dto.ForecastDto;
 import net.c7j.weather.geomagnetic.dao.entity.ForecastEntity;
 import net.c7j.weather.geomagnetic.mapper.impl.ForecastDtoMapper;
+import net.c7j.weather.geomagnetic.model.base.IndexType;
+import net.c7j.weather.geomagnetic.model.dto.ForecastDto;
 import net.c7j.weather.geomagnetic.test.helper.GeneratorHelper;
 import net.c7j.weather.geomagnetic.test.tag.IntegrationTest;
 import org.junit.jupiter.api.BeforeEach;
