+++ geomagnetic-api/src/test/java/net/c7j/weather/geomagnetic/mapper/TxtForecastDtoMapperIntegrationTest.java
@@ -1,10 +1,10 @@
 package net.c7j.weather.geomagnetic.mapper;
 
-import net.c7j.weather.geomagnetic.dao.base.IndexType;
-import net.c7j.weather.geomagnetic.dao.base.TimeIntervalType;
-import net.c7j.weather.geomagnetic.dao.dto.TxtForecastDto;
 import net.c7j.weather.geomagnetic.dao.entity.ForecastEntity;
 import net.c7j.weather.geomagnetic.mapper.impl.TxtForecastDtoMapper;
+import net.c7j.weather.geomagnetic.model.base.IndexType;
+import net.c7j.weather.geomagnetic.model.base.TimeIntervalType;
+import net.c7j.weather.geomagnetic.model.dto.TxtForecastDto;
 import net.c7j.weather.geomagnetic.test.helper.GeneratorHelper;
 import net.c7j.weather.geomagnetic.test.tag.IntegrationTest;
 import org.junit.jupiter.api.BeforeEach;
