+++ geomagnetic-api/src/test/java/net/c7j/weather/geomagnetic/test/helper/AssertionHelper.java
@@ -2,8 +2,8 @@ package net.c7j.weather.geomagnetic.test.helper;
 
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
-import net.c7j.weather.geomagnetic.dao.dto.ForecastDto;
-import net.c7j.weather.geomagnetic.dao.dto.ForecastResponse;
+import net.c7j.weather.geomagnetic.model.dto.ForecastDto;
+import net.c7j.weather.geomagnetic.model.dto.ForecastResponse;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
